Use generic codec for annotations
